Refactor OrganizationSetttings
